refactor: API 레이어를 src/api/로 이동 및 auth 훅 콜로케이션
