+++ src/app/auth/hooks/useSignupAgreements.ts
